+++ Add-Debugger.ps1
@@ -1,5 +1,5 @@
 <#PSScriptInfo
-.VERSION 1.2.0
+.VERSION 2.0.0
 .AUTHOR Roman Kuzmin
 .COPYRIGHT (c) Roman Kuzmin
 .TAGS Debug
@@ -35,14 +35,17 @@
 		Specifies the output file used instead of Out-Host.
 		A separate console is opened for watching its tail.
 
+		Do not let this file to grow too large.
+		Invoke `new` when watching gets slower.
+
 .Parameter Context
-		Specifies the number of context source lines to show.
+		Specifies the default numbers of context source lines to show.
+		One or two integers, line numbers before and after the current.
 
-.Parameter XPos
-		Horizontal position of the input box.
-
-.Parameter YPos
-		Vertical position of the input box.
+.Parameter Environment
+		Specifies the user environment variable name for keeping the state.
+		It is also used as the input box title. The state includes context
+		line numbers and input box coordinates.
 
 .Parameter ReadHost
 		Tells to use Read-Host instead of the input box.
@@ -77,11 +80,15 @@
 #>
 
 param(
-	[Parameter()]
-	[string]$Path,
-	[int]$Context,
-	[int]$XPos = -1,
-	[int]$YPos = -1,
+	[Parameter(Position=0)]
+	[string]$Path
+	,
+	[ValidateCount(1, 2)]
+	[ValidateRange(0, 9999)]
+	[int[]]$Context = @(0, 0)
+	,
+	[string]$Environment
+	,
 	[switch]$ReadHost
 )
 
@@ -92,7 +99,7 @@ if (Test-Path Variable:\_Debugger) {
 
 # Removes and gets debugger handlers.
 function global:Remove-Debugger {
-	$instance = [System.Management.Automation.Runspaces.Runspace]::DefaultRunspace.Debugger
+	$instance = [runspace]::DefaultRunspace.Debugger
 	$type = [System.Management.Automation.Debugger]
 	$e = $type.GetEvent('DebuggerStop')
 	$v = $type.GetField('DebuggerStop', ([System.Reflection.BindingFlags]'NonPublic, Instance')).GetValue($instance)
@@ -113,25 +120,61 @@ function global:Restore-Debugger {
 	$null = Remove-Debugger
 	if ($_Debugger.Handlers) {
 		foreach($handler in $_Debugger.Handlers) {
-			[System.Management.Automation.Runspaces.Runspace]::DefaultRunspace.Debugger.add_DebuggerStop($handler)
+			[runspace]::DefaultRunspace.Debugger.add_DebuggerStop($handler)
 		}
 	}
 	Remove-Variable _Debugger -Scope Global -Force
 }
 
+function global:Read-DebuggerState {
+	param($Environment, $Context)
+	$n, $m = $Context
+	if ($null -eq $m) {
+		$m = $n
+	}
+	$x = -1
+	$y = -1
+	if ($Environment) {
+		if ($string = [Environment]::GetEnvironmentVariable($Environment, 'User')) {
+			try {
+				$b = New-Object System.Data.Common.DbConnectionStringBuilder
+				$b.set_ConnectionString($string)
+				$n = [int]$b['n']
+				$m = [int]$b['m']
+				$x = [int]$b['x']
+				$y = [int]$b['y']
+			}
+			catch {}
+		}
+	}
+	[pscustomobject]@{n=$n; m=$m; x=$x; y=$y}
+}
+
+function global:Save-DebuggerState {
+	if ($_Debugger.Environment) {
+		$_ = $_Debugger.State
+		$string = "n=$($_.n);m=$($_.m);x=$($_.x);y=$($_.y)"
+		[Environment]::SetEnvironmentVariable($_Debugger.Environment, $string, 'User')
+	}
+}
+
 ### Set debugger data.
 $null = New-Variable -Name _Debugger -Scope Global -Description Add-Debugger.ps1 -Option ReadOnly -Value @{
 	Path = $null
-	Context = $Context
-	XPos = [ref]$XPos
-	YPos = [ref]$YPos
+	Environment = $Environment
+	State = Read-DebuggerState $Environment $Context
 	Args = $null
 	Watch = $false
+	Console = $null
 	History = [System.Collections.ArrayList]@()
 	Handlers = Remove-Debugger
-	RefContext = [ref]0
 	Action = '?'
 	Match = $null
+	REIndent1 = [regex]'^(\s*)'
+	REIndent2 = [regex]'^(\s+)(.*)'
+	RECommand = [regex]'^\s*\$(\w+)\s*=(.*)'
+	REContext = [regex]'^(\+)?\s*(\d+)\s*(\d+)?\s*$'
+	UseAnsi = $PSVersionTable.PSVersion -ge ([Version]'7.2')
 }
 
 ### Define how to write debugger output.
@@ -140,10 +183,17 @@ if ($Path) {
 	$_Debugger.Watch = $true
 	function global:Write-Debugger {
 		param($Data)
-		[System.IO.File]::AppendAllText($_Debugger.Path, ($Data | Out-String))
+		if ($_Debugger.UseAnsi) {
+			$OutputRendering = $PSStyle.OutputRendering
+			$PSStyle.OutputRendering = 'ANSI'
+		}
+		$Data | Out-File -LiteralPath $_Debugger.Path -Encoding utf8 -ErrorAction 0 -Append
+		if ($_Debugger.UseAnsi) {
+			$PSStyle.OutputRendering = $OutputRendering
+		}
 		if ($_Debugger.Watch) {
 			$_Debugger.Watch = $false
-			Watch-Debugger $_Debugger.Path
+			Watch-Debugger
 		}
 	}
 }
@@ -164,13 +214,15 @@ if ($ReadHost) {
 else {
 	function global:Read-Debugger {
 		param($Prompt, $Default)
-		Read-InputBox $Prompt Add-Debugger $Default Step Continue $_Debugger.XPos $_Debugger.YPos
+		$title = if ($_Debugger.Environment) {$_Debugger.Environment} else {'Add-Debugger'}
+		Read-InputBox $Prompt $title $Default Step Continue $_Debugger.State
+		Save-DebuggerState
 	}
 }
 
 # Gets an input string from a dialog.
 function global:Read-InputBox {
-	param($Prompt, $Title, $Default, $Text1, $Text2, $XPos, $YPos)
+	param($Prompt, $Title, $Default, $Text1, $Text2, $State)
 
 	Add-Type -AssemblyName System.Windows.Forms, System.Drawing
 
@@ -179,9 +231,9 @@ function global:Read-InputBox {
 	$form.TopMost = $true
 	$form.Size = New-Object System.Drawing.Size(400, 132)
 	$form.FormBorderStyle = 'FixedDialog'
-	if ($XPos -and $YPos -and $XPos.Value -ge 0 -and $YPos.Value -ge 0) {
+	if ($State -and $State.x -ge 0 -and $State.y -ge 0) {
 		$form.StartPosition = 'Manual'
-		$form.Location = New-Object System.Drawing.Point($XPos.Value, $YPos.Value)
+		$form.Location = New-Object System.Drawing.Point($State.x, $State.y)
 	}
 	else {
 		$form.StartPosition = 'CenterScreen'
@@ -222,9 +274,9 @@ function global:Read-InputBox {
 
 	$result = $form.ShowDialog()
 
-	if ($XPos -and $YPos) {
-		$XPos.Value = $form.Location.X
-		$YPos.Value = $form.Location.Y
+	if ($State) {
+		$State.x = [Math]::Max(0, $form.Location.X)
+		$State.y = [Math]::Max(0, $form.Location.Y)
 	}
 
 	if ($result -eq 'OK') {
@@ -234,30 +286,35 @@ function global:Read-InputBox {
 
 # Starts an external file viewer.
 function global:Watch-Debugger {
-	param($Path)
-	$Path = $Path.Replace("'", "''")
+	if (($last = $_Debugger.Console) -and !$last.HasExited) {
+		try { $last.Kill() } catch {}
+	}
+	$path = $_Debugger.Path.Replace("'", "''")
 	$app = if ($PSVersionTable.PSEdition -eq 'Core' -and (Get-Command pwsh -ErrorAction 0)) {'pwsh'} else {'powershell'}
-	Start-Process $app "-NoProfile -Command `$Host.UI.RawUI.WindowTitle = 'Debug output'; Get-Content -LiteralPath '$Path' -Encoding UTF8 -Wait"
+	$_Debugger.Console = Start-Process $app "-NoProfile -Command `$Host.UI.RawUI.WindowTitle = 'Debug output'; Get-Content -LiteralPath '$path' -Encoding UTF8 -Wait" -PassThru
 }
 
 # Writes the current invocation info.
 function global:Write-DebuggerInfo {
-	param($InvocationInfo, $Context)
+	param($InvocationInfo, $State)
 
 	# write position message
 	if ($_ = $InvocationInfo.PositionMessage) {
 		Write-Debugger ($_.Trim())
 	}
 
-	# done?
+	if (!$State.n -and !$State.m) {
+		return
+	}
+
 	$file = $InvocationInfo.ScriptName
-	if ($Context -le 0 -or !$file -or !(Test-Path -LiteralPath $file)) {
+	if (!$file -or !(Test-Path -LiteralPath $file)) {
 		return
 	}
 
 	# write file lines
 	$markIndex = $InvocationInfo.ScriptLineNumber - 1
-	Write-DebuggerFile $file ($markIndex - $Context) (2 * $Context + 1) $markIndex
+	Write-DebuggerFile $file ($markIndex - $State.n) ($State.n + 1 + $State.m) $markIndex
 }
 
 # Writes the specified file lines.
@@ -274,17 +331,18 @@ function global:Write-DebuggerFile {
 	$lines = @(Get-Content -LiteralPath $Path -TotalCount ($LineIndex + $LineCount) -Force -ErrorAction 0)
 
 	# leading spaces
-	$space = ($lines[$LineIndex .. -1] | .{process{
-		if ($_ -match '^(\s*)\S') {
-			($matches[1] -replace "`t", '    ').Length
-		}
+	$re = $_Debugger.REIndent1
+	$indent = ($lines[$LineIndex .. -1] | .{process{
+		$re.Match($_).Groups[1].Value.Replace("`t", '    ').Length
 	}} | Measure-Object -Minimum).Minimum
 
 	# write lines with a mark
 	Write-Debugger ''
+	$re = $_Debugger.REIndent2
 	do {
-		if (($line = $lines[$LineIndex]) -match '^(\s*)(\S.*)') {
-			$line = ($matches[1] -replace "`t", '    ').Substring($space) + $matches[2]
+		$line = $lines[$LineIndex]
+		if (($m = $re.Match($line)).Success) {
+			$line = $m.Groups[1].Value.Replace("`t", '    ').Substring($indent) + $m.Groups[2].Value
 		}
 		$mark = if ($LineIndex -eq $MarkIndex) {'>>'} else {'  '}
 		Write-Debugger ('{0,4}:{1} {2}' -f ($LineIndex + 1), $mark, $line)
@@ -294,7 +352,7 @@ function global:Write-DebuggerFile {
 }
 
 ### Add DebuggerStop handler.
-[System.Management.Automation.Runspaces.Runspace]::DefaultRunspace.Debugger.add_DebuggerStop({
+[runspace]::DefaultRunspace.Debugger.add_DebuggerStop({
 	# write breakpoints
 	if ($_.Breakpoints) {&{
 		Write-Debugger ''
@@ -309,7 +367,7 @@ function global:Write-DebuggerFile {
 	}}
 
 	# write debug location
-	Write-DebuggerInfo $_.InvocationInfo $_Debugger.Context
+	Write-DebuggerInfo $_.InvocationInfo $_Debugger.State
 	Write-Debugger ''
 
 	# restore parent $_
@@ -394,23 +452,26 @@ function global:Write-DebuggerFile {
 		}
 
 		### <number>
-		if ([int]::TryParse($_Debugger.Action, $_Debugger.RefContext)) {
-			Write-DebuggerInfo $_Debugger.Args.InvocationInfo $_Debugger.RefContext.Value
-			if ($_Debugger.Action[0] -eq "+") {
-				$_Debugger.Context = $_Debugger.RefContext.Value
+		if ($_Debugger.REContext.IsMatch($_Debugger.Action)) {
+			&{
+				$m = $_Debugger.REContext.Match($_Debugger.Action)
+				$n1 = [int]$m.Groups[2].Value
+				$n2 = [int]$(if ($m.Groups[3].Success) {$m.Groups[3].Value} else {$n1})
+				Write-DebuggerInfo $_Debugger.Args.InvocationInfo @{n=$n1; m=$n2}
+				if ($m.Groups[1].Success) {
+					$_Debugger.State.n = $n1
+					$_Debugger.State.m = $n2
+					Save-DebuggerState
+				}
 			}
-			$_Debugger.Action = $_Debugger.LastAction
 			continue
 		}
 
-		### watch
-		if ('w' -eq $_Debugger.Action) {
-			if ($_Debugger.Path) {
-				Watch-Debugger $_Debugger.Path
-			}
-			else {
-				Write-Debugger 'Debugger output file is not used.'
-			}
+		### new
+		if ('new' -eq $_Debugger.Action -and $_Debugger.Path) {
+			Remove-Item -LiteralPath $_Debugger.Path
+			Write-Debugger (Get-Date)
+			Watch-Debugger
 			continue
 		}
 
@@ -430,14 +491,14 @@ function global:Write-DebuggerFile {
 				'  k            Write call stack (Get-PSCallStack).'
 				'  K            Write detailed call stack using Format-List.'
 				''
-				'  <n>          Write debug location in context of <n> lines.'
-				'  +<n>         Set location context preference to <n> lines.'
-				'  k <s> <n>    Write source at stack <s> in context of <n> lines.'
+				'  n1 [n2]      Write debug location in context of n lines.'
+				'  + n1 [n2]    Set location context preference to n lines.'
+				'  k s n1 [n2]  Write source at stack s in context of n lines.'
 				''
+				'  r            Write last commands invoked on debugging.'
 				if ($_Debugger.Path) {
-				'  w            Restart watching the debugger output file.'
+				'  new          Remove output file and start watching new.'
 				}
-				'  r            Write last PowerShell commands invoked on debugging.'
 				'  <empty>      Repeat the last command if it was StepInto, StepOver.'
 				'  <command>    Invoke any PowerShell <command> and write its output.'
 				''
@@ -445,30 +506,37 @@ function global:Write-DebuggerFile {
 			continue
 		}
 
-		### stack <s> <n>
+		### stack s n1 [n2]
 		Set-Alias k debug_stack
-		function debug_stack([Parameter()][int]$s, [int]$n) {
+		function debug_stack([Parameter()][int]$s, $n1, $n2) {
 			$stack = @(Get-PSCallStack)
 			if ($s -ge $stack.Count) {
 				Write-Debugger 'Out of range of the call stack.'
 				return
 			}
-			$1 = $stack[$s]
-			if (!($file = $1.ScriptName)) {
+			$stack = $stack[$s]
+			if (!($file = $stack.ScriptName)) {
 				Write-Debugger 'The caller has no script file.'
 				return
 			}
-			if ($n -le 0) {$n = 5}
-			$markIndex = $1.ScriptLineNumber - 1
+			if ($null -eq $n1) {
+				$n1 = 5
+				$n2 = 5
+			}
+			else {
+				$n1 = [Math]::Max(0, [int]$n1)
+				$n2 = [Math]::Max(0, $(if ($null -eq $n2) {$n1} else {[int]$n2}))
+			}
+			$markIndex = $stack.ScriptLineNumber - 1
 			Write-Debugger $file
-			Write-DebuggerFile $file ($markIndex - $n) (2 * $n + 1) $markIndex
+			Write-DebuggerFile $file ($markIndex - $n1) ($n1 + 1 + $n2) $markIndex
 		}
 
 		### invoke command
 		try {
 			$_Debugger.History.Remove($_Debugger.Action)
 			$null = $_Debugger.History.Add($_Debugger.Action)
-			$_Debugger.Match = [regex]::Match($_Debugger.Action, '^\s*\$(\w+)\s*=(.*)')
+			$_Debugger.Match = $_Debugger.RECommand.Match($_Debugger.Action)
 			if ($_Debugger.Match.Success) {
 				$value = . ([scriptblock]::Create($_Debugger.Match.Groups[2]))
 				Set-Variable -Name ($_Debugger.Match.Groups[1]) -Value $value -Scope 1